[BE-33] refactor: Authority 폴더 구조 변경
