+++ src/main/java/com/chatty/service/user/UserService.java
@@ -4,7 +4,7 @@ import com.chatty.constants.Code;
 import com.chatty.dto.user.request.*;
 import com.chatty.dto.user.response.UserResponse;
 import com.chatty.dto.user.response.UserResponseDto;
-import com.chatty.entity.user.Authority;
+import com.chatty.constants.Authority;
 import com.chatty.entity.user.User;
 import com.chatty.exception.CustomException;
 import com.chatty.jwt.JwtTokenProvider;
